# internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: tileserve-go
  description: >
    A small Go HTTP server for managing maps and their versioned tile uploads,
    backed by Postgres and secured with JWT authentication. Maps have
    versioned zip uploads extracted to `<data-root>/<uuid>/<version>/` on
    disk, served back at `/maps/{id}/version/{version}/...`.
  version: "1.0.0"
servers:
  - url: http://localhost:8085
...
security:
  - bearerAuth: []
  - tokenQuery: []
paths:
...
  /login/oidc:
    get:
      tags: [auth]
      summary: Start OpenID Connect login
      description: >-
        Only present when OIDC is configured (see GET /auth/methods).
        Redirects the browser to the configured provider's authorization
        endpoint to begin the login flow; the provider redirects back to
        GET /login/oidc/callback on completion.
      security: []
      parameters:
        - name: redirect
          in: query
          required: false
          description: >-
            Same-origin absolute path to redirect back to after login,
            defaulting to /ui/. Any other value (a full URL, a
            protocol-relative path, etc.) is rejected in favor of the
            default.
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the provider's authorization endpoint
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Every bearer token is a JWT, of one of two kinds: a human login/
        refresh token (`POST /login`, `POST /refresh`), signed HS256 with a
        server-wide shared secret and carrying no `kid` header; or an API
        key token, signed RS256 by the caller with the private half of a
        key pair whose public half was registered via `POST
        /users/{username}/api-keys`, carrying a `kid` header naming that
        key's id. API key tokens must be short-lived (no more than 15
        minutes between `iat` and `exp`) — the caller mints and signs a
        fresh one for each use rather than reusing a single long-lived
        token.
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: JWT (login or API key) passed as a query parameter instead of the Authorization header.
...